# Splits feature_combinations into multiple functions