--- a/R/features.R
+++ b/R/features.R
@@ -6,8 +6,8 @@
 #' The returned data.table contains the following columns
 #' \describe{
 #' \item{ID}{Positive integer. Unique key for combination}
-#' \item{features}{List}
-#' \item{nfeautres}{Positive integer}
+#' \item{features}{List of integer vectors}
+#' \item{nfeatures}{Positive integer}
 #' \item{N}{Positive integer}
 #' }
 #'