--- a/man/shapley_weights.Rd
+++ b/man/shapley_weights.Rd
@@ -4,7 +4,7 @@
 \alias{shapley_weights}
 \title{Calculate Shapley weight}
 \usage{
-shapley_weights(m, N, s, weight_zero_m)
+shapley_weights(m, N, s, weight_zero_m = 10^6)
 }
 \arguments{
 \item{m}{Integer. Total number of features}