… NAMESPACE
… man/observation_weights.Rd
… tests/testthat/test-features.R