#### Summarized 
- Extract parts of `feature_combinations` into two new functions called `feature_exact` and `feature_not_exact`. All the functions are stored in `R/features.R`
- Removes `observation_weights` since it is not used anywhere in the codebase. The function was previously stored in `R/shapley.R`.  
- Updates `shapley_weigths` so that it handle cases with zero features and m features, i.e. where `s = 0` or `s = m`.
- Moves sampling of features into `sampling_features_cpp`. This part will only be used if `exact = FALSE` in `feature_combinations()`.
- Adds tests for functions in `R/features.R`. See `tests/testthat/test-features.R`.

Fixes  #71 and #72. These issues came up when writing the tests.

#### Benchmarks

Ran the following chunk using the version from `master` and `nikolai/tests_features`
```
new <- bench::mark(
  new = feature_combinations(m = 15, exact = FALSE, noSamp = 1e4), 
  iterations = 50, 
  filter_gc = FALSE
)
new <- data.table::data.table(new)
knitr::kable(new[, .SD, .SDcols = 1:10])
```
#### Current
```
|expression |   min|  mean| median|   max|  itr/sec| mem_alloc| n_gc| n_itr| total_time|
|:----------|-----:|-----:|------:|-----:|--------:|---------:|----:|-----:|----------:|
|new        | 129ms| 149ms|  144ms| 365ms| 6.704927|        0B|   19|    50|      7.46s|
```
#### Master
```
|expression |   min|  mean| median|   max| itr/sec| mem_alloc| n_gc| n_itr| total_time|
|:----------|-----:|-----:|------:|-----:|-------:|---------:|----:|-----:|----------:|
|old        | 652ms| 770ms|  751ms| 1.03s| 1.29869|        0B|  185|    50|      38.5s|
```